+++ src/std_components/cores/simple_core.hpp
@@ -1,5 +1,5 @@
-#ifndef SINUCA3_SIMPLE_MEMORY_HPP_
-#define SINUCA3_SIMPLE_MEMORY_HPP_
+#ifndef SINUCA3_SIMPLE_CORE_HPP_
+#define SINUCA3_SIMPLE_CORE_HPP_
 
 //
 // Copyright (C) 2024  HiPES - Universidade Federal do Paraná
@@ -19,32 +19,39 @@
 //
 
 /**
- * @file simple_memory.hpp
- * @details Public API of the SimpleMemory: component for SiNUCA3 which just
- * responds immediatly for every request. I.e., the perfect memory: big and
- * works at the light speed!
+ * @file simple_core.hpp
+ * @details Public API of the SimpleCore, a testing core that executes
+ * everything in a single clock cycle.
  */
 
-#include "../sinuca3.hpp"
+#include "../../sinuca3.hpp"
 
 /**
- * @details SimpleMemory is a MemoryComponent that just responds immediatly for
- * every request. I.e., it's the perfect memory: big and works at the light
- * speed!
+ * @details SimpleCore executes everything in a single cycle. You can optionally
+ * set an instructionMemory and a dataMemory pointers to components that extend
+ * Component<MemoryPacket>.
  */
-class SimpleMemory : public sinuca::Component<sinuca::MemoryPacket> {
+class SimpleCore : public sinuca::Component<sinuca::InstructionPacket> {
   private:
-    unsigned long numberOfRequests;
+    sinuca::Component<sinuca::MemoryPacket>* instructionMemory;
+    sinuca::Component<sinuca::MemoryPacket>* dataMemory;
+
+    int instructionConnectionID;
+    int dataConnectionID;
+    unsigned long numFetchedInstructions;
 
   public:
-    inline SimpleMemory() : numberOfRequests(0) {};
+    inline SimpleCore()
+        : instructionMemory(NULL),
+          dataMemory(NULL),
+          numFetchedInstructions(0) {}
     virtual int FinishSetup();
     virtual int SetConfigParameter(const char* parameter,
                                    sinuca::config::ConfigValue value);
     virtual void Clock();
     virtual void Flush();
     virtual void PrintStatistics();
-    ~SimpleMemory();
+    ~SimpleCore();
 };
 
-#endif  // SINUCA3_SIMPLE_MEMORY_HPP_
+#endif  // SINUCA3_SIMPLE_CORE_HPP_
