+++ src/std_components/memory/simple_memory.cpp
@@ -22,7 +22,7 @@
 
 #include "simple_memory.hpp"
 
-#include "../utils/logging.hpp"
+#include "../../utils/logging.hpp"
 
 int SimpleMemory::SetConfigParameter(const char* parameter,
                                      sinuca::config::ConfigValue value) {
