+++ src/std_components/predictors/ras.cpp
@@ -22,7 +22,7 @@
 
 #include "ras.hpp"
 
-#include "../utils/logging.hpp"
+#include "../../utils/logging.hpp"
 
 int Ras::FinishSetup() {
     if (this->size == 0) {
