+++ src/std_components/std_components.cpp
@@ -24,9 +24,11 @@
 #include <cstring>
 
 #include "../sinuca3.hpp"
+#include "cores/simple_core.hpp"
 #include "engine_debug_component.hpp"
-#include "simple_core.hpp"
-#include "simple_memory.hpp"
+#include "memory/simple_memory.hpp"
+#include "predictors/interleavedBTB.hpp"
+#include "predictors/ras.hpp"
 
 using namespace sinuca;
 
@@ -37,6 +39,8 @@ engine::Linkable* sinuca::CreateDefaultComponentByClass(const char* name) {
 
     COMPONENT(SimpleMemory);
     COMPONENT(SimpleCore);
+    COMPONENT(Ras);
+    COMPONENT(BranchTargetBuffer);
 
     return NULL;
 }
